调整部分通达信系统批量数量限制的最大数类型

+++ quotes/stock_security_bars.go
@@ -9,7 +9,7 @@ import (
 )
 
 const (
-	TDX_SECURITY_BARS_MAX = uint16(800) // 单次最大获取800条K线数据
+	TDX_SECURITY_BARS_MAX = 800 // 单次最大获取800条K线数据
 )
 
 // SecurityBars K线
