+++ quotes/stock_security_quotes.go
@@ -8,7 +8,7 @@ import (
 )
 
 const (
-	TDX_SECURITY_QUOTES_MAX = uint16(80) // 单次最大获取80条实时数据
+	TDX_SECURITY_QUOTES_MAX = 80 // 单次最大获取80条实时数据
 )
 
 // SecurityQuotesPackage 盘口五档报价
